+++ cercle-3/fdf/srcs/gui/panel/panel.c
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/22 14:28:37 by abdoali           #+#    #+#             */
-/*   Updated: 2025/12/13 17:00:26 by abdoali          ###   ########.fr       */
+/*   Updated: 2025/12/21 00:33:24 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ static void	draw_rect(t_gui *gui, t_vec2 pos, t_vec2 size, int color)
 	if (!img->img_addr)
 		return ;
 	y = pos.y;
-	while (y < pos.y + size.y && y < gui->window->height)
+		while (y < pos.y + size.y && y < (int)gui->window->height)
 	{
 		x = pos.x;
 		while (x < pos.x + size.x && x < GUI_PANEL_WIDTH)
